tlog-tiles: add a title
This seems to be the only C2SP document without one. Add one so it's
easier to cite. I went with "Tiled Transparency Logs" because that's how
the document seems to refer to this ("A tiled transparency log is
defined by ...").

The other tlog documents are named like "Transparency Log Cosignatures",
which would suggest "Transparency Log Tiles", but the document is the
entrypoint for the whole tlog API, so "Tiled Transparency Logs" seemed
more appropriate.