tlog-mirror: first draft of the spec
As part of this, this updates tlog-cosignature to cover cosigners with
more expanded guarantees.

This incorporates some feedback since
https://github.com/C2SP/C2SP/issues/133, both on the issue and
out-of-band:

- tlog-ext-cosignature is replaced with cosignature/v2

- The background pull-based update process is replaced with push-based API

One piece of feedback I have not yet incorporated is that maybe
cosignature/v2 is unnecessary and instead we should instead just expand
the semantics of the existing cosignature/v1 signature (somewhat of a
sleight of hand, but should be compatible?), and live with the cosigner
name not being committed by having a MUST-level requirement against key
reuse. Happy to do either, and I figure we can use the PR to discuss
this.

Another thing to call out is I'm currently citing
draft-davidben-tls-merkle-tree-certs-05 for some "subtree" machinery.
It's a stable reference, but I'm not sure if that's better respecified
for now.