tlog-cosignature: be a bit more correct w.r.t. leap seconds
Leap seconds are the worst. I believe, formally, one needs to explicitly
exclude them for a description of a POSIX timestamp to be correct. An
example in HTTP:
https://www.rfc-editor.org/rfc/rfc9651#name-dates

(Or you do what POSIX does and actually define it as a mathematical
expression of the time's UTC name, and just leave unspecified how you
relate that back to an actual time.)
https://pubs.opengroup.org/onlinepubs/9799919799/basedefs/V1_chap04.html#tag_04_19