tlog-cosignature: Allow additional statements
This change is motivated by the need to support witnesses that assert
additional statements, e.g., "I also mirrored all log entries" and "I
also reproduced the sparse Merkle tree root in the extension lines".

This was suggested as cosignature/v2 by davidben and lukevalenta.  The
main change would have been to tie the witness name to the cosignature
such that the name could be used as a cosigning context.  This is a good
idea and we will consider it in a future cosignature/v2.  Until then, we
require *distinct keys* to establish such separate cosigning contexts.

Thanks: mhutchinson and eliasrudberg for suggesting text improvements.