tlog-tiles: handle a tree size of zero in pruning section (#155)
* Reword text around minimum index in tlog-tiles.md

Reword minimum index text to clarify that the minimum index is a lower
bound, but the actual entry at the minimum index might not be available
(e.g., if the minimum index is equal to the tree size).

Also clarify that we can always compute the root hash of the tree of
size _equal_ to (not just strictly greater than) the minimum index.  To
see this, say the minimum index is X.

* If X=0, the root hash is just SHA256(<empty string>)
* If X=tree_size, the root hash can be computed from the 'right edge' of
  nodes (or pulled from the current checkpoint).
* If 0<X<tree_size, the inclusion proof of the index-X element in the
  tree of size X+1 necessarily includes the tree nodes to compute the
  root hash for the tree of size X.

* Use 'greater than or equal to' consistently